chore: simplify pages directory by moving simple pages out of their own directories
