+++ pages/partner.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next'
-import styles from '../../styles/Home.module.css'
+import styles from '../styles/Home.module.css'
 
 const Partner: NextPage = () => {
   return (
